app/requirements.txt
altair
pandas
streamlit
skops
scikit-learn
mlflow>=2,<3
dagshub
python-dotenv